Renamed containers to projects
